hardware_interface cpp integration

+++ ros2_ws/src/okmr_hardware_interface/okmr_hardware_interface/include/thrust_to_pwm.h
@@ -7,5 +7,5 @@ void updateThrustTableIfVoltageChanged(double voltage);
 
 
 double thrustToNormalizedPWM(double thrust);  // use if you want to input the normalized values between -1.0 and 1.0
-int thrustToPWM(double thrust);               // use for 1100–1900 µs
+double thrustToPWM(double thrust);            // use for 1100–1900 µs
